refactor(model-zoom): rename to model-cartesian-charts
